requirements.txt
requests
fastapi
uvicorn[standard] 
python-doctr[torch]
matplotlib
vietocr
pandas
typing
python-multipart
numpy


